refac: shared - ui 세부 분리
